+++ tests/deprecated_api/test_remove_1-7.py
@@ -112,7 +112,6 @@ def test_v1_7_0_moved_get_progress_bar_dict(tmpdir):
 
     trainer = Trainer(
         default_root_dir=tmpdir,
-        progress_bar_refresh_rate=None,
         fast_dev_run=True,
     )
     test_model = TestModel()
@@ -129,11 +128,6 @@ def test_v1_7_0_moved_get_progress_bar_dict(tmpdir):
 def test_v1_7_0_trainer_prepare_data_per_node(tmpdir):
     with pytest.deprecated_call(match="Setting `prepare_data_per_node` with the trainer flag is deprecated in v1.5.0"):
         _ = Trainer(prepare_data_per_node=False)
-
-
-def test_v1_7_0_stochastic_weight_avg_trainer_constructor(tmpdir):
-    with pytest.deprecated_call(match=r"Setting `Trainer\(stochastic_weight_avg=True\)` is deprecated in v1.5"):
-        _ = Trainer(stochastic_weight_avg=True)
 
 
 @pytest.mark.parametrize("terminate_on_nan", [True, False])
@@ -255,11 +249,6 @@ def test_v1_7_0_deprecate_add_get_queue(tmpdir):
 
     with pytest.deprecated_call(match=r"`LightningModule.get_from_queue` method was deprecated in v1.5"):
         trainer.fit(model)
-
-
-def test_v1_7_0_progress_bar_refresh_rate_trainer_constructor(tmpdir):
-    with pytest.deprecated_call(match=r"Setting `Trainer\(progress_bar_refresh_rate=1\)` is deprecated in v1.5"):
-        _ = Trainer(progress_bar_refresh_rate=1)
 
 
 def test_v1_7_0_lightning_logger_base_close(tmpdir):